requirements.txt
annotated-types==0.7.0
anyio==3.7.1
astroid==3.2.3
async-generator==1.10
attrs==23.2.0
bcrypt==4.2.0
black==24.4.2
cached-property==1.5.2
certifi==2024.7.4
cffi==1.16.0
click==8.1.7
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
dill==0.3.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-login==1.10.1
fonttools==4.53.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==2.0.1
opencv-python==4.10.0.84
orjson==3.10.6
outcome==1.3.0.post0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
pyparsing==3.1.2
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.31
sqlalchemy-orm==1.2.10
starlette==0.37.2
tomli==2.0.1
tomlkit==0.13.0
trio==0.21.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
